requirements.txt
# Core
pydantic>=2.0.0
click>=8.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
loguru>=0.7.0

# AI/ML
ollama>=0.1.7
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Utils
psutil>=5.9.0
rich>=13.0.0

# Dev
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
